added more pods
eplicas refer to the number of identical copies or instances of a Pod or workload that should be running at any given time. Replicas are a fundamental concept in Kubernetes and are used to ensure high availability, scalability, and reliability of applications. Replicas are typically managed using controllers like Deployment, StatefulSet, ReplicaSet, and ReplicationController.
$kubectl apply -f k8s/nginx/deployment.yaml
$kubectl get pods